Using the openapi spec for generation with reverse compatability layer
