+++ asana/resources/batch_api.py
@@ -1,0 +1,6 @@
+
+from .gen.batch_api import _BatchAPI
+
+class BatchAPI(_BatchAPI):
+    """BatchAPI resource"""
+    pass
