+++ asana/resources/gen/project_statuses.py
@@ -1,62 +1,45 @@
 
 class _ProjectStatuses:
-    """A _project status_ is an update on the progress of a particular project, and is sent out to all project
-    followers when created. These updates include both text describing the update and a color code intended to
-    represent the overall state of the project: "green" for projects that are on track, "yellow" for projects
-    at risk, and "red" for projects that are behind.
-    
-    Project statuses can be created and deleted, but not modified.
-    """
 
     def __init__(self, client=None):
         self.client = client
-  
-    def create_in_project(self, project, params={}, **options): 
-        """Creates a new status update on the project.
-        
-        Returns the full record of the newly created project status update.
 
-        Parameters
-        ----------
-        project : {Gid} The project on which to create a status update.
-        [data] : {Object} Data for the request
-          - text : {String} The text of the project status update.
-          - color : {String} The color to associate with the status update. Must be one of `"red"`, `"yellow"`, or `"green"`.
-        """
-        path = "/projects/%s/project_statuses" % (project)
-        return self.client.post(path, params, **options)
-        
-    def find_by_project(self, project, params={}, **options): 
-        """Returns the compact project status update records for all updates on the project.
-
-        Parameters
-        ----------
-        project : {Gid} The project to find status updates for.
+    def create_project_status(self, project_gid, params={}, **options):
+        """Create a project status
+        :param str project_gid: Globally unique identifier for the project. (required)
         [params] : {Object} Parameters for the request
+        :return: ProjectStatusResponse
         """
-        path = "/projects/%s/project_statuses" % (project)
-        return self.client.get_collection(path, params, **options)
-        
-    def find_by_id(self, project_status, params={}, **options): 
-        """Returns the complete record for a single status update.
-
-        Parameters
-        ----------
-        project-status : {Gid} The project status update to get.
-        [params] : {Object} Parameters for the request
-        """
-        path = "/project_statuses/%s" % (project_status)
+        path = "/projects/{project_gid}/project_statuses".replace("project_gid", project_gid)
         return self.client.get(path, params, **options)
-        
-    def delete(self, project_status, params={}, **options): 
-        """Deletes a specific, existing project status update.
-        
-        Returns an empty data record.
 
-        Parameters
-        ----------
-        project-status : {Gid} The project status update to delete.
+
+    def delete_project_status(self, project_status_path_gid, params={}, **options):
+        """Delete a project status
+        :param str project_status_path_gid: The project status update to get. (required)
+        [params] : {Object} Parameters for the request
+        :return: EmptyObject
         """
-        path = "/project_statuses/%s" % (project_status)
-        return self.client.delete(path, params, **options)
-        
+        path = "/project_statuses/{project_status_gid}".replace("project_status_path_gid", project_status_path_gid)
+        return self.client.get(path, params, **options)
+
+
+    def get_project_status(self, project_status_path_gid, params={}, **options):
+        """Get a project status
+        :param str project_status_path_gid: The project status update to get. (required)
+        [params] : {Object} Parameters for the request
+        :return: ProjectStatusResponse
+        """
+        path = "/project_statuses/{project_status_gid}".replace("project_status_path_gid", project_status_path_gid)
+        return self.client.get(path, params, **options)
+
+
+    def get_project_statuses(self, project_gid, params={}, **options):
+        """Get statuses from a project
+        :param str project_gid: Globally unique identifier for the project. (required)
+        [params] : {Object} Parameters for the request
+        :return: list[ProjectStatusCompact]
+        """
+        path = "/projects/{project_gid}/project_statuses".replace("project_gid", project_gid)
+        return self.client.get(path, params, **options)
+
