+++ asana/resources/gen/stories.py
@@ -1,75 +1,55 @@
 
 class _Stories:
-    """A _story_ represents an activity associated with an object in the Asana
-    system. Stories are generated by the system whenever users take actions such
-    as creating or assigning tasks, or moving tasks between projects. _Comments_
-    are also a form of user-generated story.
-    """
 
     def __init__(self, client=None):
         self.client = client
-  
-    def find_by_task(self, task, params={}, **options): 
-        """Returns the compact records for all stories on the task.
 
-        Parameters
-        ----------
-        task : {Gid} Globally unique identifier for the task.
+    def create_story(self, task_gid, params={}, **options):
+        """Create a story on a task
+        :param str task_gid: The task to operate on. (required)
         [params] : {Object} Parameters for the request
+        :return: StoryResponse
         """
-        path = "/tasks/%s/stories" % (task)
-        return self.client.get_collection(path, params, **options)
-        
-    def find_by_id(self, story, params={}, **options): 
-        """Returns the full record for a single story.
-
-        Parameters
-        ----------
-        story : {Gid} Globally unique identifier for the story.
-        [params] : {Object} Parameters for the request
-        """
-        path = "/stories/%s" % (story)
+        path = "/tasks/{task_gid}/stories".replace("task_gid", task_gid)
         return self.client.get(path, params, **options)
-        
-    def create_on_task(self, task, params={}, **options): 
-        """Adds a comment to a task. The comment will be authored by the
-        currently authenticated user, and timestamped when the server receives
-        the request.
-        
-        Returns the full record for the new story added to the task.
 
-        Parameters
-        ----------
-        task : {Id} Globally unique identifier for the task.
-        [data] : {Object} Data for the request
-          - text : {String} The plain text of the comment to add.
-        """
-        path = "/tasks/%s/stories" % (task)
-        return self.client.post(path, params, **options)
-        
-    def update(self, story, params={}, **options): 
-        """Updates the story and returns the full record for the updated story.
-        Only comment stories can have their text updated, and only comment stories and
-        attachment stories can be pinned. Only one of `text` and `html_text` can be specified.
 
-        Parameters
-        ----------
-        story : {Gid} Globally unique identifier for the story.
-        [data] : {Object} Data for the request
-          - [text] : {String} The plain text with which to update the comment.
-          - [html_text] : {String} The rich text with which to update the comment.
-          - [is_pinned] : {Boolean} Whether the story should be pinned on the resource.
+    def delete_story(self, story_gid, params={}, **options):
+        """Delete a story
+        :param str story_gid: Globally unique identifier for the story. (required)
+        [params] : {Object} Parameters for the request
+        :return: EmptyObject
         """
-        path = "/stories/%s" % (story)
-        return self.client.put(path, params, **options)
-        
-    def delete(self, story, params={}, **options): 
-        """Deletes a story. A user can only delete stories they have created. Returns an empty data record.
+        path = "/stories/{story_gid}".replace("story_gid", story_gid)
+        return self.client.get(path, params, **options)
 
-        Parameters
-        ----------
-        story : {Gid} Globally unique identifier for the story.
+
+    def get_stories_for_task(self, task_gid, params={}, **options):
+        """Get stories from a task
+        :param str task_gid: The task to operate on. (required)
+        [params] : {Object} Parameters for the request
+        :return: list[StoryCompact]
         """
-        path = "/stories/%s" % (story)
-        return self.client.delete(path, params, **options)
-        
+        path = "/tasks/{task_gid}/stories".replace("task_gid", task_gid)
+        return self.client.get(path, params, **options)
+
+
+    def get_story(self, story_gid, params={}, **options):
+        """Get a story
+        :param str story_gid: Globally unique identifier for the story. (required)
+        [params] : {Object} Parameters for the request
+        :return: StoryResponse
+        """
+        path = "/stories/{story_gid}".replace("story_gid", story_gid)
+        return self.client.get(path, params, **options)
+
+
+    def update_story(self, story_gid, params={}, **options):
+        """Update a story
+        :param str story_gid: Globally unique identifier for the story. (required)
+        [params] : {Object} Parameters for the request
+        :return: StoryResponse
+        """
+        path = "/stories/{story_gid}".replace("story_gid", story_gid)
+        return self.client.get(path, params, **options)
+
