+++ asana/resources/typeahead.py
@@ -1,0 +1,6 @@
+
+from .gen.typeahead import _Typeahead
+
+class Typeahead(_Typeahead):
+    """Typeahead resource"""
+    pass
