+++ asana/resources/workspace_memberships.py
@@ -1,0 +1,6 @@
+
+from .gen.workspace_memberships import _WorkspaceMemberships
+
+class WorkspaceMemberships(_WorkspaceMemberships):
+    """WorkspaceMemberships resource"""
+    pass
